fix: Fix element-template-generator-maven-plugin for long version values (#7373)
* Initial plan

* Fix element-template-generator to support long version numbers

- Change VersionedElementTemplate.version from Integer to Long
- Change ElementTemplate.version from int to long
- Update getBasicElementTemplate to return Optional and skip unparseable templates
- Templates with version > Integer.MAX_VALUE (like Web Modeler's epoch timestamps) now work

Agent-Logs-Url: https://github.com/camunda/connectors/sessions/c1d29290-f1dc-4856-84b6-21609227951c


* fix: Claude forgot to change an int into a Long

* fix: Support long version values in element-template-generator

- Changed version field from int/Integer to long/Long throughout the codebase
- Updated ElementTemplate annotation to accept long version values
- Updated ElementTemplate record, builder, and related classes
- Updated PropertyGroup and CommonProperties to use Long
- Updated GeneratorConfiguration to use Long for templateVersion
- Added logging when skipping unparseable template files in maven plugin
- Added test coverage for long version values
- Fixes compatibility with Web Modeler OpenAPI-generated templates that use Unix epoch timestamps

Agent-Logs-Url: https://github.com/camunda/connectors/sessions/514b81bf-c6eb-4434-ac55-fa1c98814811


* fix: small fix in a test

---------

+++ element-template-generator/annotations/src/main/java/io/camunda/connector/generator/java/annotation/ElementTemplate.java
@@ -54,7 +54,7 @@ public @interface ElementTemplate {
    * <p><b>NB</B>: It is recommended to specify the version explicitly for element templates used in
    * production.
    */
-  int version() default 0;
+  long version() default 0;
 
   /**
    * Link to the documentation page for the connector. Will be used by the Camunda Modeler to
